🎨(lib-classroom) separate tests modules from main package
@faker-js/faker is a big package, it weights a lot in the bundle size
of our apps, and we only use it for tests.
In order to optimize the bundle size, we separate the tests modules
from the main package.
