+++ src/frontend/apps/lti_site/apps/classroom/components/SelectContent/SelectContentResource/index.spec.tsx
@@ -1,7 +1,7 @@
 import { screen, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import fetchMock from 'fetch-mock';
-import { classroomMockFactory } from 'lib-classroom';
+import { classroomMockFactory } from 'lib-classroom/tests';
 import { playlistMockFactory } from 'lib-components/tests';
 import { render } from 'lib-tests';
 import React from 'react';
