+++ src/frontend/apps/lti_site/apps/classroom/components/SelectContent/SelectContentSection/index.spec.tsx
@@ -1,5 +1,5 @@
 import { screen } from '@testing-library/react';
-import { classroomMockFactory } from 'lib-classroom';
+import { classroomMockFactory } from 'lib-classroom/tests';
 import { render } from 'lib-tests';
 import React from 'react';
