+++ src/frontend/packages/lib_classroom/src/components/ClassroomInfoBar/index.spec.tsx
@@ -5,7 +5,7 @@ import { report } from 'lib-components';
 import { render } from 'lib-tests';
 import { Settings } from 'luxon';
 
-import { classroomMockFactory } from '@lib-classroom/utils/tests/factories';
+import { classroomMockFactory } from '@lib-classroom/tests/factories';
 import { wrapInClassroom } from '@lib-classroom/utils/wrapInClassroom';
 
 import { ClassroomInfoBar } from '.';
