+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/index.spec.tsx
@@ -6,7 +6,7 @@ import { render } from 'lib-tests';
 import { DateTime } from 'luxon';
 import React from 'react';
 
-import { classroomMockFactory } from '@lib-classroom/utils';
+import { classroomMockFactory } from '@lib-classroom/tests/factories';
 import { wrapInClassroom } from '@lib-classroom/utils/wrapInClassroom';
 
 import { ClassroomWidgetProvider } from '.';
