+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/DeleteClassroom/index.spec.tsx
@@ -8,7 +8,7 @@ import {
 } from 'lib-components';
 import { render } from 'lib-tests';
 
-import { classroomMockFactory } from '@lib-classroom/utils';
+import { classroomMockFactory } from '@lib-classroom/tests/factories';
 import { wrapInClassroom } from '@lib-classroom/utils/wrapInClassroom';
 
 import { DeleteClassroom } from '.';
