+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/Recordings/Recording/DeleteClassroomRecordingButton/index.spec.tsx
@@ -8,7 +8,7 @@ import { DateTime } from 'luxon';
 import {
   classroomMockFactory,
   classroomRecordingMockFactory,
-} from '@lib-classroom/utils';
+} from '@lib-classroom/tests/factories';
 
 import { DeleteClassroomRecordingButton } from '.';
