+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/Recordings/Recording/index.spec.tsx
@@ -9,7 +9,7 @@ import {
   classroomMockFactory,
   classroomRecordingMockFactory,
   classroomRecordingVodMockFactory,
-} from '@lib-classroom/utils/tests/factories';
+} from '@lib-classroom/tests/factories';
 
 import { Recording } from '.';
