+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/Recordings/index.spec.tsx
@@ -13,7 +13,7 @@ import React from 'react';
 import {
   classroomMockFactory,
   classroomRecordingMockFactory,
-} from '@lib-classroom/utils/tests/factories';
+} from '@lib-classroom/tests/factories';
 import { wrapInClassroom } from '@lib-classroom/utils/wrapInClassroom';
 
 import { Recordings } from '.';
