+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/SupportSharing/index.spec.tsx
@@ -5,7 +5,7 @@ import { render } from 'lib-tests';
 import { Settings } from 'luxon';
 import React from 'react';
 
-import { classroomMockFactory } from '@lib-classroom/utils/tests/factories';
+import { classroomMockFactory } from '@lib-classroom/tests/factories';
 import { wrapInClassroom } from '@lib-classroom/utils/wrapInClassroom';
 
 import { SupportSharing } from '.';
