+++ src/frontend/packages/lib_classroom/src/components/DashboardClassroomForm/index.spec.tsx
@@ -7,7 +7,7 @@ import { Deferred, render } from 'lib-tests';
 import { Settings } from 'luxon';
 import React from 'react';
 
-import { classroomMockFactory } from '@lib-classroom/utils/tests/factories';
+import { classroomMockFactory } from '@lib-classroom/tests/factories';
 
 import DashboardClassroomForm from '.';
