+++ src/frontend/packages/lib_classroom/src/data/queries/index.spec.tsx
@@ -6,7 +6,7 @@ import { WrapperReactQuery } from 'lib-tests';
 import {
   classroomDocumentMockFactory,
   classroomMockFactory,
-} from '@lib-classroom/utils/tests/factories';
+} from '@lib-classroom/tests/factories';
 
 import {
   useClassroom,
