+++ src/frontend/packages/lib_classroom/src/tests/index.ts
@@ -1,0 +1,1 @@
+export * from './factories';
