+++ src/frontend/packages/lib_classroom/src/utils/index.ts
@@ -1,1 +1,1 @@
-export * from './tests/factories';
+export * from './wrapInClassroom';
