+++ src/frontend/packages/lib_classroom/src/utils/wrapInClassroom.tsx
@@ -1,6 +1,7 @@
-import { CurrentClassroomProvider } from 'hooks';
 import { Classroom } from 'lib-components';
 import React, { ReactNode } from 'react';
+
+import { CurrentClassroomProvider } from '@lib-classroom/hooks';
 
 export const wrapInClassroom = (component: ReactNode, classroom: Classroom) => {
   return (
